chore(report-details): :truck: Move component files to appropriate folder
