+++ app/routes/reports_.$slug.tsx
@@ -3,11 +3,11 @@ import { LoaderFunction, json } from "@remix-run/node";
 import { Link, MetaArgs, MetaFunction, useLoaderData } from "@remix-run/react";
 import parse from "html-react-parser";
 import { ChevronLeft, MapPin } from "lucide-react";
-import DynamicCategoryIcon from "~/components/dynamic-category-icon";
-import FundingProgress from "~/components/funding-progress";
-import ReportSidebar from "~/components/report-sidebar";
-import ReportSupportFeed from "~/components/report-support-feed";
+import FundingProgress from "~/components/report-details/funding-progress";
+import ReportSidebar from "~/components/report-details/report-sidebar";
+import ReportSupportFeed from "~/components/report-details/report-support-feed";
 import { Badge } from "~/components/ui/badge";
+import { DynamicCategoryIcon } from "~/components/ui/dynamic-category-icon";
 import { Separator } from "~/components/ui/separator";
 import { fetchReportBySlug } from "~/impact-reports.server";
 import { Report } from "~/types";
